Update dependencies and refactor chat components for improved performance
- Added @radix-ui/react-progress to package dependencies for enhanced UI feedback.
- Commented out unused components and code in various chat-related files to streamline the codebase.
- Updated file upload handling and removed deprecated socket utility files to reduce complexity.
- Refactored chat components for better performance and dynamic imports.
- Cleaned up unused imports and optimized component structure across multiple files.
