+++ apps/web/app/(main)/new/page.tsx
@@ -1,13 +1,13 @@
-import NewChatArea from "@/components/chat/NewChatArea";
+// import NewChatArea from "@/components/chat/NewChatArea";
 import ChatInfoPage from "@/components/newlook/chat-info";
-import ChatSidebar from "@/components/sidebar/NewSidebar";
+// import ChatSidebar from "@/components/sidebar/NewSidebar";
 
 export default function ChatPage() {
   return (
     <div className="flex h-screen">
-      <ChatSidebar />
+      {/* <ChatSidebar /> */}
       <div className="flex-1">
-        <NewChatArea chatName="John Doe" isTyping={true}  />
+        {/* <NewChatArea chatName="John Doe" /> */}
       </div>
       {/* <ChatInfoPage /> */}
     </div>
